move messages to their own dir
