+++ crates/p2p/src/message/hardware_challenge.rs
